Bug 1502969 - Set last-watershed version to 60.0 for win64. r+a=jorgk
There were no win64 builds prior to 60.0, looking for one earlier than that
causes errors during release-update-verify-config essentially halting the
release process.

A similar fix was committed to c-c in rev ec9dbe9537af. The other changes
in that revision to not apply to c-esr60 as by-release-type is not available
in 60.xESR.

Differential Revision: https://phabricator.services.mozilla.com/D10098