Bug 1557102 - Work around inline styles being ignored. r=darktrojan a=jorgk
For some unknown reason, inline styles in overlays are ignored until something
triggers re-evaluation of the element style. This causes inline styles to be
invisible whenever an add-on uses inline styles, but does not alter the styled
element using code.

To work around the issue, this commit forces styles to be evaluated by
setting the 'display' style to the already evaluated value. That forces any
inline styles to become visible immediately.