Bug 1621357 - [release] Fix scopes and environment variables in partials jobs. r+a=mkmelin
This just removes the scopes and environment variables from the task before
submitting it to Taskcluster. funsize.py itself does not need to have these
set in order to function.

While fixing the Datadog scopes, I noticed that the MAR channel environment
variable would not get set for Thunderbird. Since this patch is intended to
run only on comm-esr68 and will never go elsewhere it's hardcoded.