Bug 1650041 - When adding a new address book or list, only change the UI if the change was made with the UI r=pmorris
Previously, the UI would detect something was new and change to it immediately. With this patch,
that will not happen. Instead, the dialogs for creating new books and lists return the URI of the
new object, and the change happens when the dialog closes.

This patch finally removes the comments "we can optimize this later" added 9 years ago. I guess
now is later.

Differential Revision: https://phabricator.services.mozilla.com/D82314