Bug 1991002 - Remove redundant nsIMsgIncomingServer.supportsDiskSpace. r=mkmelin
nsIMsgIncomingServer.supportsDiskSpace was only used by AccountManager.js
as part of the logic to decide if the am-offline panel should be shown or
not. And IMIncomingServer was the only server type which returned false.
So, much clearer to just remove all the boilerplate and do an explicit
server.type check in AccountManager.js instead.

Differential Revision: https://phabricator.services.mozilla.com/D267132