Bug 1984116 - Capture telemetry for EWS accounts in Glean. r=edicharry,tobyp a=coreycb
I went through a few iterations on this patch, including recording telemetry on
each request, or during server setup. Ultimately, these approaches were a bit
naff because the telemetry data would be skewed by the account setup flow
(wherein we create a new `nsIMsgIncomingServer` for each configuration we test).
I settled with triggering the reporting at the same place as other somewhat
related telemetry reporting, which we only perform once at startup (or shortly
after). This means new accounts aren't recorded until after a restart, but
that's probably fine, and it seems in line with how we record other telemetry
data.

The actual telemetry recording is still happening in Rust-land, because that's
where the logic for reading server versions lives. I went a bit back-and-forth
on how to make it usable from JavaScript code, and settled on a stateless
method on `IEwsClient`.

Differential Revision: https://phabricator.services.mozilla.com/D268278