Bug 1998342 - Display a notification when a calendar has connection issues. r=mkmelin
I've built a connection notifications module with the intention of using it for other account types
at a later date (replacing alertHook.sys.mjs). Pulling out the actual network status codes to an
appropriate place to use them was a bit of a mess but I think it's the right thing to do.

Differential Revision: https://phabricator.services.mozilla.com/D271535