Bug 1971144 - Prevent messages.deleteAttachments() to modify/destroy encrypted messages. r=mkmelin
The WebExtension messages API can act on the original message, or on the decrypted
message (the decrypted message is not stored), when calling `messages.getFull()`
or `messages.getRaw()`. Functions listing attachments of a message by default act
on the decrypted message only, without an option to list attachments of the
original encrypted message (this may be added later).

Problematic is the function to delete attachments. It was assumed to also acts on
the decrypted message. This however is not supported by Thunderbird and the function
actually acts on the original message, potentially deleting the encrypted container
attachment, if the provided `partName` matched.

Thunderbird disables all UI entry points to delete attachments of encrypted messages.
This patch updates the API documentation and also disallows deleting attachments of
encrypted messages.

The related Thunderbird bug is tracked here:
https://bugzilla.mozilla.org/show_bug.cgi?id=288700

Differential Revision: https://phabricator.services.mozilla.com/D271403