Bug 1973284 - Persistently store the id of space buttons in associated tabs. r=mkmelin
To restore the association between an extension space (button) and a re-opened
space tab, the ids of associated spaces need to be stored persistently. The
persistent extension storage in `nativeTabInfo._ext` is ideal, since data is
stored only if really required and automatically thrown away if the tab is closed.

Differential Revision: https://phabricator.services.mozilla.com/D271741