Bug 1975854 - Fix race condition when multiple extensions register a listener for the compose.onAfterSend event. r=mkmelin
The compose window may be destroyed while the event result is still being
prepared asynchronously, which could cause some listeners to not notify
their handlers.

This patch introduces the `onConvertSync` callback, which is executed while
the compose window is guaranteed to exist. Because this callback must be
called from inside `MsgOperationObserver`, the calls to the other callbacks
(`onSuccess` and `onFailure`) have also been moved there.

Differential Revision: https://phabricator.services.mozilla.com/D272863